requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3

pytest==8.0.0
pytest-django==4.8.0
